Added celery tasks

+++ api/models.py
@@ -10,6 +10,6 @@ class Submission(models.Model):
         WRONG = 'wrong'
 
     uuid = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-    internal_id = models.IntegerField(unique=True, null=True, blank=True)
+    internal_id = models.BigIntegerField(unique=True, null=True, blank=True)
     function_code = models.TextField()
     status = models.TextField(choices=StatusEnum.choices, default=StatusEnum.EVALUATION)
